requirements.txt
django
djangorestframework
bs4
requests
django-star-ratings
pytz
pillow
lxml
